[BPK-4276] Restructure remaining C pages
