+++ docs/src/pages/CheckboxPage/AndroidCheckboxPage.js
