+++ docs/src/pages/CheckboxPage/WebCheckboxPage.js
