+++ docs/src/pages/ChipsPage/AndroidChipPage.js
